pyproject.toml
[project]
name = "finetune"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bitsandbytes>=0.48.2",
    "colorama>=0.4.6",
    "datasets>=4.4.1",
    "docling>=2.63.0",
    "huggingface-hub>=0.36.0",
    "litellm>=1.80.7",
    "peft>=0.18.0",
    "python-dotenv>=1.0.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
    "trl>=0.25.1",
]


# Configure the PyTorch index for CUDA 12.4 (Adjust if you need 11.8 or 12.6)
[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu130" }]
torchvision = [{ index = "pytorch-cu130" }]
